refactoring example
