+++ example/cookbook/apps.py
@@ -3,5 +3,5 @@ from django.apps import AppConfig
 
 class CookbookConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
+    label = 'example_cookbook'
     name = 'example.cookbook'
-    label = 'cookbook'
