+++ example/templates/speech_to_text/card/ai_detail_card.html
